Remove the link from provider statement to profiles.
Issue #949: Give providers the ability to "edit and resubmit" applications
